Q2LAL layout test with PEX
This is an extension of a test currently documented at https://revcomp.org/make-a-chip/. The .mag files represent a Magic/Sky130 layout of a Q2LAL stage. The .spice file includes parasitics (but should be extractable from the .mag files). The function of the layout is the same as is simulated in sQ2.cir but the layout demanded a different hierarchy, so the file sQ130 is a substantial rewrite of sQ2.cir. The parasitics can be turned on/off with ".if (0)"-type lines. Dissipation per clock is 1.99532E-14 and 6.02372E-15 J with parasitics on and off.